requirements.txt
# AWS Dependencies
boto3>=1.26.0
sagemaker>=2.140.0

# Data Science Dependencies
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
joblib>=1.2.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# API Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Development
pytest>=7.2.0
black>=23.0.0
python-dotenv>=1.0.0

# Excel support for pandas
openpyxl>=3.1.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Additional AWS dependencies
botocore>=1.29.0
